feat: Transaksi Pembayaran Sekolah ke Invoice (#117)
* feat: transaksi to invoice 🔥

* invoice dari spp tidak bisa diedit ❌

* fix: total amount di invoice ✅

* remove unused lines ❌

* fix addToInvoice 🔥

* fix: revisi dari Pak Diar

* fix: message error

* feat: autoNumeric 🔥

* chore: autoNumeric invoice detail 🔥

* wip: autoNumeric ❌

* formatAngka 🔥

* void invoice 🔥

* remove unused lines ❌

+++ app/Http/Controllers/InvoiceController.php
@@ -2,9 +2,9 @@
 
 namespace App\Http\Controllers;
 
+use App\Actions\Invoice\CreateNewInvoiceNumber;
 use App\Http\Requests\InvoiceRequest;
 use App\Models\Invoice;
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
 
@@ -33,13 +33,13 @@ class InvoiceController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(InvoiceRequest $request)
+    public function store(InvoiceRequest $request, CreateNewInvoiceNumber $createNewInvoiceNumber)
     {
         DB::beginTransaction();
         try {
             $invoice = new Invoice();
             $invoice->school_id = session('school_id');
-            $invoice->invoice_number = $request->invoice_number;
+            $invoice->invoice_number = $request->invoice_number ?? $createNewInvoiceNumber->generate();
             $invoice->note = $request->note;
             $invoice->invoice_date = $request->invoice_date;
             $invoice->due_date = $request->due_date;
@@ -58,19 +58,19 @@ class InvoiceController extends Controller
     }
 
     /**
-     * Display the specified resource.
-     */
-    public function show(Invoice $invoice)
-    {
-        //
-    }
-
-    /**
      * Show the form for editing the specified resource.
      */
     public function edit(Invoice $invoice)
     {
         if ($invoice->school_id != session('school_id')) abort(404);
+
+        // cek status dan kembalikan jika statusnya sudah PUBLISHED
+        if ($invoice->is_original == false)
+            return redirect()->back()->withToastError('Ups! Invoice tidak berhak diubah.');
+
+        // cek status dan kembalikan jika statusnya bukan DRAFT
+        if ($invoice->is_posted != Invoice::POSTED_DRAFT)
+            return redirect()->back()->withToastError('Ups! Invoice tidak berhak untuk diubah.');
 
         $data['title'] = "Ubah " . $this->title;
         $data['invoice'] = $invoice;
@@ -80,9 +80,17 @@ class InvoiceController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Invoice $invoice)
+    public function update(InvoiceRequest $request, Invoice $invoice)
     {
         if ($invoice->school_id != session('school_id')) abort(404);
+
+        // cek status dan kembalikan jika statusnya sudah PUBLISHED
+        if ($invoice->is_original == false)
+            return redirect()->back()->withToastError('Ups! Invoice tidak berhak diubah.');
+
+        // cek status dan kembalikan jika statusnya bukan DRAFT
+        if ($invoice->is_posted != Invoice::POSTED_DRAFT)
+            return redirect()->back()->withToastError('Ups! Invoice tidak berhak untuk diubah.');
 
         DB::beginTransaction();
         try {
@@ -112,14 +120,24 @@ class InvoiceController extends Controller
     {
         if ($invoice->school_id != session('school_id')) abort(404);
 
+        // cek status dan kembalikan jika statusnya bukan DRAFT
+        if ($invoice->is_posted != Invoice::POSTED_DRAFT)
+            return response()->json([
+                'msg' => 'Ups! Invoice tidak berhak untuk dihapus.'
+            ], 500);
+
         DB::beginTransaction();
         try {
             $invoice->delete();
             DB::commit();
         } catch (\Throwable $th) {
             DB::rollBack();
-            return redirect()->back()->withToastError('Ups! ' . $th->getMessage());
+            return response()->json([
+                'msg' => 'Ups! ' . $th->getMessage()
+            ], 500);
         }
-        return to_route('invoices.index')->withToastSuccess('Berhasil menghapus invoice!');
+        return response()->json([
+            'msg' => 'Berhasil menghapus invoice!'
+        ], 200);
     }
 }
